// src/api/schemas/reporteVentasAgenciaResponseDataItemsItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.16
 */

/**
 * Detalle de cada venta
 */
export type ReporteVentasAgenciaResponseDataItemsItem = {
  id?: string;
  voucher?: string;
  fecha_venta?: string;
  agencia_id?: string;
  agencia_nombre?: string;
  agente_id?: string;
  agente_nombre?: string;
  precio_venta?: number;
  costo_procesamiento?: number;
  base_comision?: number;
  porcentaje?: string;
  monto?: string;
};